# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/subscription": {
            "get": {
                "description": "Get subscriptions by user_id or sub name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Get subscriptions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id in UUID string format",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "subscription name",
                        "name": "subscription_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.GetSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new subscription",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "Add new subscription",
                "parameters": [
                    {
                        "description": "SubscriptionPayload payload",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.AddSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.AddSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/aggregate": {
            "get": {
                "description": "get sum of specified subscriptions for a user in certain period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "get aggregate data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "subscription name",
                        "name": "subscription",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-02",
                        "description": "Min starting date of subscription (YYYY-MM-DD)",
                        "name": "from_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-02",
                        "description": "Min starting date of subscription (YYYY-MM-DD)",
                        "name": "to_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.SubscriptionSumResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscription/{id}": {
            "put": {
                "description": "update subscription data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "update subscription data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subscription id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.UpdateSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete subscription by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscription"
                ],
                "summary": "delete subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subscription id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.DeleteSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/objects.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "objects.AddSubscriptionRequest": {
            "type": "object",
            "properties": {
                "subscription": {
                    "$ref": "#/definitions/objects.SubscriptionPayload"
                }
            }
        },
        "objects.AddSubscriptionResponse": {
            "type": "object"
        },
        "objects.DeleteSubscriptionResponse": {
            "type": "object"
        },
        "objects.ErrorResponse": {
            "type": "object"
        },
        "objects.GetSubscriptionResponse": {
            "type": "object",
            "properties": {
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/objects.SubscriptionPayload"
                    }
                }
            }
        },
        "objects.SubscriptionPayload": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "11-2021"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "yandex"
                },
                "start_date": {
                    "type": "string",
                    "example": "10-2020"
                },
                "user_id": {
                    "type": "string",
                    "example": "019e5b65-04a0-7833-bd67-856ea7d20900"
                }
            }
        },
        "objects.SubscriptionSumResponse": {
            "type": "object",
            "properties": {
                "from_date": {
                    "type": "string",
                    "example": "2020-10-01"
                },
                "subscription": {
                    "type": "string",
                    "example": "yandex"
                },
                "sum": {
                    "type": "integer",
                    "example": 10
                },
                "to_date": {
                    "type": "string",
                    "example": "2021-11-01"
                },
                "user_id": {
                    "type": "string",
                    "example": "019e5b65-04a0-7833-bd67-856ea7d20900"
                }
            }
        },
        "objects.UpdateSubscriptionResponse": {
            "type": "object"
        }
    }
}